Completed

+++ src/app/modules/products/product.service.ts
@@ -12,18 +12,40 @@ const getAllProductFromDB = async () => {
 };
 
 const getSingleProductFromDB = async (productId: string) => {
+  if (!productId) {
+    throw new Error("Product id is required");
+  }
   const result = await Product.findById(productId);
+  if (!result) {
+    throw new Error("Product not found");
+  }
   return result;
 };
 
 const updateSingleProductFromDB = async (productId: string, data: IProduct) => {
+  if (!productId) {
+    throw new Error("Product id is required");
+  }
+
+  if (!data) {
+    throw new Error("Product data is required");
+  }
+
   const result = await Product.findByIdAndUpdate(productId, data, {
     new: true,
   });
+
+  if (!result) {
+    throw new Error("Product not found");
+  }
+
   return result;
 };
 
 const deleteSingleProductFromDB = async (productId: string) => {
+  if (!productId) {
+    throw new Error("Product id is required");
+  }
   const result = await Product.findByIdAndDelete(productId);
 
   if (!result) {
